+++ src/serialize.js
@@ -5,53 +5,52 @@ import safeJsonValue from 'safe-json-value'
 
 import { setConstructorArgs } from './args.js'
 import { safeListKeys } from './check.js'
+import { applyList } from './list.js'
 import { listProps } from './props.js'
 import { applyTransformObject } from './transform.js'
 
 // Serialize error instances into plain objects deeply
-export const serializeDeep = (value, transformObject, parents) => {
+export const serializeDeep = (value, options, parents) => {
   const parentsA = [...parents, value]
 
   if (!isErrorInstance(value)) {
-    return serializeRecurse(value, transformObject, parentsA)
+    return serializeRecurse(value, options, parentsA)
   }
 
   const error = normalizeException(value)
   const errorObject = serializeError(error)
-  const errorObjectA = serializeRecurse(errorObject, transformObject, parentsA)
-  applyTransformObject(transformObject, errorObjectA, error)
-  const errorObjectB = safeJsonValue(errorObjectA, { shallow: false }).value
-  return errorObjectB
+  const errorObjectA = serializeRecurse(errorObject, options, parentsA)
+  applyTransformObject(errorObjectA, error, options)
+  const errorObjectB = applyList(errorObjectA, options)
+  const errorObjectC = safeJsonValue(errorObjectB, { shallow: false }).value
+  return errorObjectC
 }
 
 // Serialize a possible error instance into a plain object
-export const serializeShallow = (value, transformObject) => {
+export const serializeShallow = (value, options) => {
   if (!isErrorInstance(value)) {
     return value
   }
 
   const error = normalizeException(value)
   const errorObject = serializeError(error)
-  applyTransformObject(transformObject, errorObject, error)
-  const errorObjectA = safeJsonValue(errorObject, { shallow: true }).value
-  return errorObjectA
+  applyTransformObject(errorObject, error, options)
+  const errorObjectA = applyList(errorObject, options)
+  const errorObjectB = safeJsonValue(errorObjectA, { shallow: true }).value
+  return errorObjectB
 }
 
 const serializeError = (error) =>
   Object.fromEntries([...getProps(error), ...setConstructorArgs(error)])
 
 const getProps = (error) =>
-  listProps(error)
-    .map((propName) => [propName, error[propName]])
-    .filter(hasValue)
+  listProps(error).map((propName) => [propName, error[propName]])
 
-const hasValue = ([, value]) => value !== undefined
-
-const serializeRecurse = (value, transformObject, parents) => {
+const serializeRecurse = (value, options, parents) => {
   if (Array.isArray(value)) {
     return value
       .filter((child) => !parents.includes(child))
-      .map((child) => serializeDeep(child, transformObject, parents))
+      .map((child) => serializeDeep(child, options, parents))
   }
 
   if (isPlainObj(value)) {
@@ -60,7 +59,7 @@ const serializeRecurse = (value, transformObject, parents) => {
         .filter((propName) => !parents.includes(value[propName]))
         .map((propName) => [
           propName,
-          serializeDeep(value[propName], transformObject, parents),
+          serializeDeep(value[propName], options, parents),
         ]),
     )
   }